requirements.txt
django
djangorestframework
fpdf2